requirements.txt
pytest
typer[all]
rich
playwright
google-generativeai
groq